api/requirements.txt
flask
flask-cors
torch
torchvision
Pillow
numpy
